refactor: changing the place of the doCall function to optimize code organization.
