+++ packages/gateway/src/gateWay.ts
@@ -15,7 +15,6 @@
  *
  */
 import * as ccipread from "@chainlink/ccip-read-server";
-import { Interface } from "ethers/lib/utils";
 
 // Application Binary Interfaces
 export const abi: string[] = [
@@ -25,46 +24,6 @@ export const abi: string[] = [
   "function addr(bytes32 node, uint coinType) public view returns(bytes memory result)",
   "function text(bytes32 node, string key) public view returns(string memory result)",
 ];
-/**
- * Executes a function call on the specified server using the provided ABI and arguments.
- *
- * @param {ccipread.Server} server - The server instance to perform the function call.
- * @param {string[]} abi - The ABI (Application Binary Interface) array describing the function.
- * @param {string} path - The path or address to which the call is made.
- * @param {string} type - The type of the function to be called.
- * @param {any[]} args - The arguments required for the function call.
- * @throws {Error} - Throws an error if the handler for the specified function type is unknown or if the server response has a non-200 status.
- */
-async function doCall(
-  server: ccipread.Server,
-  abi: string[],
-  path: string,
-  type: string,
-  args: any[]
-) {
-  const iface = new Interface(abi);
-  const handler = server.handlers[iface.getSighash(type)];
-
-  // Check if the handler for the specified function type is registered
-  if (!handler) {
-    throw Error("Unknown handler");
-  }
-  // Encode function data using ABI and arguments
-  const calldata = iface.encodeFunctionData(type, args);
-  // Make a server call with encoded function data
-  const result = await server.call({ to: path, data: calldata });
-
-  // Check if the server response has a non-200 status
-  if (result.status !== 200) {
-    throw Error(result.body.message);
-  }
-  // Decode the function result if the function has outputs, otherwise return an empty array
-  if (handler.type.outputs !== undefined) {
-    return iface.decodeFunctionResult(handler.type, result.body.data);
-  } else {
-    return [];
-  }
-}
 
 // Creating a ccip-read server
 const server = new ccipread.Server();
@@ -119,4 +78,4 @@ server.add(abi, [
 ]);
 
 // Exporting the created gateway and the function that allow communication with it.
-export { server, doCall };
+export { server };
