+++ packages/gateway/test/gateWay.spec.ts
@@ -15,8 +15,10 @@
  *
  */
 import { describe, it, beforeAll, expect } from "vitest";
-import { server, doCall, abi } from "../src/gateWay";
+import { server, abi } from "../src/gateWay";
 import ethers from "ethers";
+import * as ccipread from "@chainlink/ccip-read-server";
+import { Interface } from "ethers/lib/utils";
 
 // Defining the port where the gateway will run
 const port = 3001;
@@ -29,6 +31,47 @@ const TEST_ADDRESS = "0x1234567890123456789012345678901234567890";
 // Function to convert string into bytes32
 function createBytes32(data: string): string {
   return ethers.utils.id(data);
+}
+
+/**
+ * Executes a function call on the specified server using the provided ABI and arguments.
+ *
+ * @param {ccipread.Server} server - The server instance to perform the function call.
+ * @param {string[]} abi - The ABI (Application Binary Interface) array describing the function.
+ * @param {string} path - The path or address to which the call is made.
+ * @param {string} type - The type of the function to be called.
+ * @param {any[]} args - The arguments required for the function call.
+ * @throws {Error} - Throws an error if the handler for the specified function type is unknown or if the server response has a non-200 status.
+ */
+async function doCall(
+  server: ccipread.Server,
+  abi: string[],
+  path: string,
+  type: string,
+  args: any[]
+) {
+  const iface = new Interface(abi);
+  const handler = server.handlers[iface.getSighash(type)];
+
+  // Check if the handler for the specified function type is registered
+  if (!handler) {
+    throw Error("Unknown handler");
+  }
+  // Encode function data using ABI and arguments
+  const calldata = iface.encodeFunctionData(type, args);
+  // Make a server call with encoded function data
+  const result = await server.call({ to: path, data: calldata });
+
+  // Check if the server response has a non-200 status
+  if (result.status !== 200) {
+    throw Error(result.body.message);
+  }
+  // Decode the function result if the function has outputs, otherwise return an empty array
+  if (handler.type.outputs !== undefined) {
+    return iface.decodeFunctionResult(handler.type, result.body.data);
+  } else {
+    return [];
+  }
 }
 
 // Testing calls to the gateway
